requirements-local.txt
# For local testing on non-NVIDIA platforms (e.g., macOS)
# This file removes version pinning to allow pip to find compatible packages.

torch
torchvision
torchaudio
diffusers
transformers
accelerate
# xformers is NVIDIA-only and will not install on Mac
# xformers
einops
omegaconf
runpod
boto3
loguru
Pillow
numpy
imageio
imageio-ffmpeg
ftfy
opencv-python-headless 